# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project API",
        "description": "Project API",
        "version": null
    },
    "servers": [
        {
            "url": "http://project.local/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/groups/{groupId}": {
            "delete": {
                "tags": [
                    "Groups"
                ],
                "summary": "Deletes a group",
                "description": "This API Endpoint deletes a group.",
                "operationId": "GroupsDelete",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Gets list of all user roles",
                "description": "This API Endpoint returns lists of all user roles.",
                "operationId": "RolesList",
                "responses": {
                    "200": {
                        "description": "Successfull Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Gets a user by ID",
                "description": "This API Endpoint get a user by ID.",
                "operationId": "UsersGet",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deletes a user",
                "description": "This API Endpoint deletes a user.",
                "operationId": "UsersDelete",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/{userId}/groups/{groupId}/assign": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Assigns a user to a group",
                "description": "This API Endpoint assigns users to groups.",
                "operationId": "UsersGroupsAssign",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfull Operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Role": {
                "description": "Role Resource",
                "properties": {
                    "id": {
                        "description": "Role ID",
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "description": "Role name",
                        "type": "string",
                        "maxLength": 32,
                        "example": "Administator"
                    }
                },
                "type": "object"
            },
            "User": {
                "description": "User Resource",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 0
                    },
                    "roleId": {
                        "description": "Role ID",
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "maxLength": 64,
                        "example": "John"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "token": []
        }
    ]
}